update obc env if it is already exisiting and unsetup without deleting obc env. there's another chunk for particularly deleting obv env (i.e. make unsetup_env)

+++ makefile
@@ -2,6 +2,7 @@
 # makefile
 SHELL := /bin/bash
 brc := source `which activate` OBc
+n_brc := `which conda` env remove -y --name OBc
 
 all: OSdiffs setupR setupPython start_message
 
@@ -57,8 +58,11 @@ conda: curl
 
 env: conda
 	if [[ -z $$(conda info --envs | grep -Ee "^OBc[ ]+/") ]]; then \
-           if [[ `uname` == "Linux"  ]]; then conda-env create --file obc_envL.yml;fi &&\
-           if [[ `uname` == "Darwin" ]]; then conda-env create --file obc_envM.yml; fi;fi
+           if [[ `uname` == "Linux"  ]]; then conda-env create --file obc_envL.yml; fi &&\
+           if [[ `uname` == "Darwin" ]]; then conda-env create --file obc_envM.yml; fi ;\
+  else\
+           if [[ `uname` == "Linux"  ]]; then $(brc) && conda-env update --file obc_envL.yml; fi &&\
+           if [[ `uname` == "Darwin" ]]; then $(brc) && conda-env update --file obc_envM.yml; fi; fi
 
 OSdiffs:
 	if [[ `uname` == "Linux" ]]; then sed -i -e "s/sed -Ee/sed -re/g" get_checkLists.sh; fi
@@ -100,11 +104,16 @@ start_message:
 	echo -e "\n\nAll dependencies installed. Please run to activate:\n\n\033[1;32m    source activate OBc\n\033[0m" &&\
 	echo -e "And to deactivate:\n\n\033[1;32m    conda deactivate\n\033[0m"
 
-unsetup:
+unsetup_py: 
+	$(brc) && python3 -c "import site; print(site.getsitepackages()[0])" |awk '{print $$1"/circling_py"}' |xargs rm -rf
+
+unsetup: unsetup_py
 	chmod -x ./circling_py/*
 	chmod -x ./circling_r/*
-	python3 -c "import site; print(site.getsitepackages()[0])" |awk '{print $$1"/circling_py"}' |xargs rm -rf
 	if [[ ! -z $$(ls ./circling_py/ | grep "__pycache__") ]]; then sudo rm -rf ./circling_py/__pycache__/; fi
 	rm circling_r/AuditionBarcode.v.2.R
 	rm circling_r/SpecimenData.R
 	if [[ ! -z $$(ls ./circling_r/ | grep -e "R-e") ]]; then rm ./circling_r/*R-e; fi
+
+unsetup_env:
+	$(n_brc)
